move parse(::String, ...) to Meta module
