+++ base/meta.jl
@@ -79,4 +79,89 @@ macro lower(mod, code)
     return :(lower($(esc(mod)), $(QuoteNode(code))))
 end
 
+
+## interface to parser ##
+
+"""
+    ParseError(msg)
+
+The expression passed to the `parse` function could not be interpreted as a valid Julia
+expression.
+"""
+struct ParseError <: Exception
+    msg::AbstractString
+end
+
+"""
+    parse(str, start; greedy=true, raise=true)
+
+Parse the expression string and return an expression (which could later be passed to eval
+for execution). `start` is the index of the first character to start parsing. If `greedy` is
+`true` (default), `parse` will try to consume as much input as it can; otherwise, it will
+stop as soon as it has parsed a valid expression. Incomplete but otherwise syntactically
+valid expressions will return `Expr(:incomplete, "(error message)")`. If `raise` is `true`
+(default), syntax errors other than incomplete expressions will raise an error. If `raise`
+is `false`, `parse` will return an expression that will raise an error upon evaluation.
+
+```jldoctest
+julia> Meta.parse("x = 3, y = 5", 7)
+(:(y = 5), 13)
+
+julia> Meta.parse("x = 3, y = 5", 5)
+(:((3, y) = 5), 13)
+```
+"""
+function parse(str::AbstractString, pos::Int; greedy::Bool=true, raise::Bool=true)
+    # pos is one based byte offset.
+    # returns (expr, end_pos). expr is () in case of parse error.
+    bstr = String(str)
+    ex, pos = ccall(:jl_parse_string, Any,
+                    (Ptr{UInt8}, Csize_t, Int32, Int32),
+                    bstr, sizeof(bstr), pos-1, greedy ? 1 : 0)
+    if raise && isa(ex,Expr) && ex.head === :error
+        throw(ParseError(ex.args[1]))
+    end
+    if ex === ()
+        raise && throw(ParseError("end of input"))
+        ex = Expr(:error, "end of input")
+    end
+    return ex, pos+1 # C is zero-based, Julia is 1-based
+end
+
+"""
+    parse(str; raise=true)
+
+Parse the expression string greedily, returning a single expression. An error is thrown if
+there are additional characters after the first expression. If `raise` is `true` (default),
+syntax errors will raise an error; otherwise, `parse` will return an expression that will
+raise an error upon evaluation.
+
+```jldoctest
+julia> Meta.parse("x = 3")
+:(x = 3)
+
+julia> Meta.parse("x = ")
+:($(Expr(:incomplete, "incomplete: premature end of input")))
+
+julia> Meta.parse("1.0.2")
+ERROR: ParseError("invalid numeric constant \\\"1.0.\\\"")
+Stacktrace:
+[...]
+
+julia> Meta.parse("1.0.2"; raise = false)
+:($(Expr(:error, "invalid numeric constant \"1.0.\"")))
+```
+"""
+function parse(str::AbstractString; raise::Bool=true)
+    ex, pos = parse(str, 1, greedy=true, raise=raise)
+    if isa(ex,Expr) && ex.head === :error
+        return ex
+    end
+    if !done(str, pos)
+        raise && throw(ParseError("extra token after end of expression"))
+        return Expr(:error, "extra token after end of expression")
+    end
+    return ex
+end
+
 end # module
