+++ test/numbers.jl
@@ -1816,7 +1816,7 @@ end
     @test isa(0b00000000000000000000000000000000000000000000000000000000000000000,UInt128)
     @test isa(0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000,UInt128)
     # remove BigInt unsigned integer literals #11105
-    @test_throws ParseError parse("0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000")
+    @test_throws Meta.ParseError Meta.parse("0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000")
     @test isa(0b11111111,UInt8)
     @test isa(0b111111111,UInt16)
     @test isa(0b1111111111111111,UInt16)
@@ -1827,7 +1827,7 @@ end
     @test isa(0b11111111111111111111111111111111111111111111111111111111111111111,UInt128)
     @test isa(0b11111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111,UInt128)
     # remove BigInt unsigned integer literals #11105
-    @test_throws ParseError parse("0b111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111")
+    @test_throws Meta.ParseError Meta.parse("0b111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111111")
 end
 @testset "octal literals" begin
     @test 0o10 == 0x8
@@ -1845,7 +1845,7 @@ end
     @test isa(0o0000000000000000000000,UInt128)
     @test isa(0o000000000000000000000000000000000000000000,UInt128)
     # remove BigInt unsigned integer literals #11105
-    @test_throws ParseError parse("0o0000000000000000000000000000000000000000000")
+    @test_throws Meta.ParseError Meta.parse("0o0000000000000000000000000000000000000000000")
     @test isa(0o11,UInt8)
     @test isa(0o111,UInt8)
     @test isa(0o11111,UInt16)
@@ -1857,7 +1857,7 @@ end
     @test isa(0o111111111111111111111111111111111111111111,UInt128)
     @test isa(0o1111111111111111111111111111111111111111111,UInt128)
     # remove BigInt unsigned integer literals #11105
-    @test_throws ParseError parse("0o11111111111111111111111111111111111111111111")
+    @test_throws Meta.ParseError Meta.parse("0o11111111111111111111111111111111111111111111")
 end
 @testset "hexadecimal literals" begin
     @test isa(0x00,UInt8)
@@ -1870,7 +1870,7 @@ end
     @test isa(0x00000000000000000,UInt128)
     @test isa(0x00000000000000000000000000000000,UInt128)
     # remove BigInt unsigned integer literals #11105
-    @test_throws ParseError parse("0x000000000000000000000000000000000")
+    @test_throws Meta.ParseError Meta.parse("0x000000000000000000000000000000000")
 
     @test isa(0x11,UInt8)
     @test isa(0x111,UInt16)
@@ -1882,7 +1882,7 @@ end
     @test isa(0x11111111111111111,UInt128)
     @test isa(0x11111111111111111111111111111111,UInt128)
     # remove BigInt unsigned integer literals #11105
-    @test_throws ParseError parse("0x111111111111111111111111111111111")
+    @test_throws Meta.ParseError Meta.parse("0x111111111111111111111111111111111")
 end
 @testset "minus sign and unsigned literals" begin
     # "-" is not part of unsigned literals
@@ -1902,7 +1902,7 @@ end
     @test isa(-0x00000000000000000,UInt128)
     @test isa(-0x00000000000000000000000000000000,UInt128)
     # remove BigInt unsigned integer literals #11105
-    @test_throws ParseError parse("-0x000000000000000000000000000000000")
+    @test_throws Meta.ParseError Meta.parse("-0x000000000000000000000000000000000")
 end
 @testset "Float32 literals" begin
     @test isa(1f0,Float32)
